Improve code.

+++ src/MitternachtBot/Modules/MitternachtModule.cs
@@ -14,7 +14,6 @@ namespace Mitternacht.Modules {
 		protected          CultureInfo CultureInfo;
 
 		public readonly string ModuleTypeName;
-		public          string LowerModuleTypeName => ModuleTypeName?.ToLowerInvariant();
 
 		public StringService  Strings      { get; set; }
 		public CommandHandler CmdHandler   { get; set; }
@@ -32,72 +31,49 @@ namespace Mitternacht.Modules {
 		}
 
 		protected string GetText(string key)
-			=> Strings.GetText(LowerModuleTypeName, key, CultureInfo);
+			=> Strings.GetText(ModuleTypeName, key, CultureInfo);
 
 		protected string GetText(string key, params object[] replacements)
-			=> Strings.GetText(LowerModuleTypeName, key, CultureInfo, replacements);
+			=> Strings.GetText(ModuleTypeName, key, CultureInfo, replacements);
 
-		protected Task<IUserMessage> MessageLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendMessageAsync(text);
-		}
+		protected Task<IUserMessage> MessageLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendMessageAsync(GetText(textKey, replacements));
 
-		protected Task<IUserMessage> ReplyLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendMessageAsync($"{Context.User.Mention} {text}");
-		}
+		protected Task<IUserMessage> ReplyLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendMessageAsync($"{Context.User.Mention} {GetText(textKey, replacements)}");
 
-		protected Task<IUserMessage> ErrorLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendErrorAsync(text);
-		}
+		protected Task<IUserMessage> ErrorLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendErrorAsync(GetText(textKey, replacements));
 
-		protected Task<IUserMessage> ReplyErrorLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendErrorAsync($"{Context.User.Mention} {text}");
-		}
+		protected Task<IUserMessage> ReplyErrorLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendErrorAsync($"{Context.User.Mention} {GetText(textKey, replacements)}");
 
-		protected Task<IUserMessage> ConfirmLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendConfirmAsync(text);
-		}
+		protected Task<IUserMessage> ConfirmLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendConfirmAsync(GetText(textKey, replacements));
 
-		protected Task<IUserMessage> ReplyConfirmLocalized(string textKey, params object[] replacements) {
-			var text = GetText(textKey, replacements);
-			return Context.Channel.SendConfirmAsync($"{Context.User.Mention} {text}");
-		}
+		protected Task<IUserMessage> ReplyConfirmLocalized(string textKey, params object[] replacements)
+			=> Context.Channel.SendConfirmAsync($"{Context.User.Mention} {GetText(textKey, replacements)}");
 
 		protected async Task<string> GetUserInputAsync(ulong userId, ulong channelId) {
 			var userInputTask = new TaskCompletionSource<string>();
 			var dsc           = (DiscordSocketClient)Context.Client;
-			try {
-				dsc.MessageReceived += MessageReceived;
-
-				if((await Task.WhenAny(userInputTask.Task, Task.Delay(10000))) != userInputTask.Task) {
-					return null;
-				}
-
-				return await userInputTask.Task;
-			} finally {
-				dsc.MessageReceived -= MessageReceived;
-			}
 
 			Task MessageReceived(SocketMessage arg) {
 				var _ = Task.Run(() => {
-					if(!(arg is SocketUserMessage userMsg) ||
-						!(userMsg.Channel is ITextChannel) ||
-						userMsg.Author.Id != userId ||
-						userMsg.Channel.Id != channelId) {
-						return Task.CompletedTask;
-					}
-
-					if(userInputTask.TrySetResult(arg.Content)) {
+					if(arg is SocketUserMessage userMsg && userMsg.Channel is ITextChannel && userMsg.Author.Id == userId && userMsg.Channel.Id == channelId && userInputTask.TrySetResult(arg.Content)) {
 						userMsg.DeleteAfter(1);
 					}
-
-					return Task.CompletedTask;
 				});
+
 				return Task.CompletedTask;
+			}
+
+			try {
+				dsc.MessageReceived += MessageReceived;
+
+				return await Task.WhenAny(userInputTask.Task, Task.Delay(10000)) != userInputTask.Task ? null : await userInputTask.Task;
+			} finally {
+				dsc.MessageReceived -= MessageReceived;
 			}
 		}
 	}
